# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShelfOps API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stores/{storeId}/locations/{locationId}/configuration": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "expectedVersion",
                  "effectiveAt",
                  "active",
                  "idempotencyKey"
                ],
                "properties": {
                  "expectedVersion": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "effectiveAt": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "effectiveUntil": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "label": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "idempotencyKey": {
                    "minLength": 1,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "storeId",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "locationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "eventId",
                    "version",
                    "before",
                    "after",
                    "effectiveAt",
                    "correlationId"
                  ],
                  "properties": {
                    "eventId": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "version": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "before": {
                      "additionalProperties": true,
                      "type": "object",
                      "properties": {}
                    },
                    "after": {
                      "additionalProperties": true,
                      "type": "object",
                      "properties": {}
                    },
                    "effectiveAt": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId",
                    "fields"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation-failed"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "name",
                          "code"
                        ],
                        "properties": {
                          "name": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "authentication-required"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "forbidden"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "not-found"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "idempotency-conflict"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "description": "Server-generated identifier for support correlation.",
                          "type": "string"
                        }
                      }
                    },
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "stale-version"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "correlationId": {
                          "description": "Server-generated identifier for support correlation.",
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "temporarily-unavailable"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shelfops_session"
      }
    },
...
  }
}